+++ packages/extension-utils/src/APIErrorDisplay/APIErrorDisplay.tsx
@@ -26,7 +26,7 @@
 
 import type { FC } from 'react'
 import React from 'react'
-import type { IAPIMethods, LookerSDKError } from '@looker/sdk-rtl'
+import type { LookerSDKError } from '@looker/sdk-rtl'
 import { Heading, SpaceVertical } from '@looker/components'
 import { APIErrorContent } from './APIErrorContent'
 import { errorHeading } from './utils'
@@ -36,21 +36,6 @@ export interface APIErrorDisplayProps {
   error: LookerSDKError
   /** true to retrieve the markdown error document from the CDN and display it instead of a link */
   showDoc?: boolean
-}
-
-/**
- * this callback function used by the ErrorDoc constructor gets and truncates
- * API error documents for use within API Explorer
- * @param _sdk required as a parameter, but it's ignored in this function
- * @param url of document (either index.json or md) to fetch
- */
-export const apiErrorDisplayFetch = async (_sdk: IAPIMethods, url: string) => {
-  let result = await (await fetch(url, { mode: 'cors' })).text()
-  const stop = result.indexOf('## API Response Type')
-  if (stop > 0) {
-    result = result.substring(0, stop - 1).trim()
-  }
-  return result
 }
 
 /**
