+++ packages/extension-utils/src/APIErrorDisplay/utils.ts
@@ -24,7 +24,26 @@
 
  */
 
-import type { LookerSDKError } from '@looker/sdk-rtl'
+import type { LookerSDKError, IAPIMethods } from '@looker/sdk-rtl'
 
+/**
+ * Default the heading to 'Unknown error' if error.message is blank for any reason
+ * @param error for heading
+ */
 export const errorHeading = (error: LookerSDKError) =>
   error.message || 'Unknown error'
+
+/**
+ * this callback function used by the ErrorDoc constructor gets and truncates
+ * API error documents for use within API Explorer
+ * @param _sdk required as a parameter, but it's ignored in this function
+ * @param url of document (either index.json or md) to fetch
+ */
+export const apiErrorDisplayFetch = async (_sdk: IAPIMethods, url: string) => {
+  let result = await (await fetch(url, { mode: 'cors' })).text()
+  const stop = result.indexOf('## API Response Type')
+  if (stop > 0) {
+    result = result.substring(0, stop - 1).trim()
+  }
+  return result
+}
